Move json.lua to the correct location
Based on telescope's dev documentation (https://github.com/nvim-telescope/telescope.nvim/blob/master/developers.md),
we should structure the pluging as follows:

```
.
└── lua
    ├── plugin_name             # Your actual plugin code
    │   ├── init.lua
    │   └── some_file.lua
    └── telescope
        └── _extensions         # The underscore is significant
            └─ plugin_name.lua  # Init and register your extension
```

to minimise the risk of module collisions.
